+++ src/portit/model/action/AccountModel.java
@@ -24,6 +24,7 @@ public class AccountModel {
 	private String newPw;				//바꿀 비민번호
 	private String confirmNewPw;		//비밀번호 확인.
 	
+	private String deletePw;
 	boolean IdentityVerification;
 	
 	
@@ -31,12 +32,15 @@ public class AccountModel {
 		this.req=_req;
 		this.login_id=_login_id;
 		
-
+		
 		//request로 부터. 입력정보를 초기화...
 		this.email=req.getParameter("email");
 		this.currentPw=req.getParameter("currentPw");
 		this.newPw=req.getParameter("newPw");
 		this.confirmNewPw=req.getParameter("confirmNewPw");
+		
+		//Delte의 Request
+		this.deletePw=req.getParameter("userdeletepw");
 		
 		
 		//DAO. 싱글톤패턴으로 테스트... 오류날수도 있음.
@@ -46,44 +50,47 @@ public class AccountModel {
 	
 	
 	
-	
-	
 	//Update
-	public void alterAccount(){
-		//임의의 메일값값을 Test
-		this.IdentityVerification=dao.validationService("kin7729@naver.com","1111");;
+	//userid,userpw
+	public void alterAccount(String userid, String userpw, String newpw, String newpwcf){
 		
-		//this.IdentityVerification=dao.validationService(email,currentPw);;//본인인증.
+		//본인인증 : 로그인된 아이디와 비밀번호가 일치하는지 조회
+		this.IdentityVerification=dao.validationService(userpw);;//본인인증.
 	
-		 
 		 
 		//ID와 PW가 맞다면.
 		if(IdentityVerification==true){
 			
-			System.out.println("MOdel: 비밀번호 일치합니다.");			
+			System.out.println("Model: 비밀번호 일치합니다.");	
 			
-			//새로운 비밀번호가 일치하다면... 
-			//dao.alterAccount(Integer.parseInt(this.newPw));
+			//새로운 비밀번호가 일치하면 변경 
+			dao.alterAccount(userid,newpw, newpw);
 		}
+		
 		//ID와 PW가 틀렸다면.
 		else{
 			//경고메세지!!!
-			System.out.println("MOdel: 비밀번호 일치하지 않습니다.");
+			System.out.println("Model: 비밀번호 일치하지 않습니다.");
 		}
 	}
 	
 	
 	
 	//삭제하기.
-	public void deleteAccount(){
-		this.IdentityVerification=dao.validationService("kin7729@naver.com","1111");
+	public Boolean deleteAccount(){
+		this.IdentityVerification=dao.validationService(this.deletePw);
 		//log_id의 pw와 일치한지....
 		// dao에서 유효성검사 쿼리가 그대로 쓰인다.
 		
+		if(IdentityVerification==true){
 		dao.deleteAccount();
+		}
+		else{
+			System.out.println("deleteAccount: 비밀번호가 일치하지 않습니다.");
+		}
+	
+	return IdentityVerification;
 	}
-	
-	
 }
